requirements.txt
fastapi
uvicorn
openai
pinecone
pdfplumber
python-docx
python-dotenv
requests
